requirements.txt
playwright~=1.49.1
atlassian-python-api==3.41.19
python-dotenv==1.0.1
urllib3~=2.3.0
pytest~=8.3.4
beautifulsoup4~=4.12.3
soupsieve~=2.6
pip~=23.2.1
attrs~=25.1.0
distro~=1.9.0
retry~=0.9.2
cryptography~=43.0.3
pillow~=11.1.0
Jinja2~=3.1.5
filelock~=3.17.0
pyee~=12.0.0
trio~=0.28.0
typing_extensions~=4.12.2
click~=8.1.8
ipython~=8.31.0
certifi~=2024.12.14
h2~=4.1.0
greenlet~=3.1.1
jmespath~=1.0.1
oauthlib~=3.2.2
blinker~=1.9.0
idna~=3.10
Deprecated~=1.2.17
requests~=2.32.3
six~=1.17.0
wrapt~=1.17.2
crewai~=0.41.1
colorama~=0.4.6
